requirements.txt
torch==1.7.1
numpy==1.19.2
tqdm==4.54.1
click==7.1.2
logzero==1.6.3
scikit-learn==0.24.1
ruamel.yaml==0.16.12
gdown==3.12.2
pandas==1.1.5
nltk==3.5
gensim==3.8.3
scipy==1.5.2
transformers==4.2.1
dgl-cu110==0.6.1
mlflow==1.15.0
joblib==1.0.0
nmslib==2.1.1
sentence-transformers==1.1.1
networkx==2.5
n2==0.1.7
